+++ src/molpy/typifier/__init__.py
@@ -1,9 +1,11 @@
 from molrs.typifier import MMFFTypifier, OPLSAATypifier
 
+from .ambertools import AmberToolsTypifier
 from .atomistic import PairTypifier
 from .clp import ClpTypifier
 
 __all__ = [
+    "AmberToolsTypifier",
     "PairTypifier",
     "OPLSAATypifier",
     "ClpTypifier",